requirements.txt
python-dotenv>=1.0.0
requests>=2.25.0
python-binance>=1.0.16
pytest>=7.0.0
